implemented styled-components
