+++ client/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\App.js":"1","C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\reportWebVitals.js":"2","C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\components\\form\\form.component.jsx":"3","C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\components\\Header\\header.component.jsx":"4"},{"size":240,"mtime":1606452642709,"results":"5","hashOfConfig":"6"},{"size":362,"mtime":1606433682700,"results":"7","hashOfConfig":"6"},{"size":393,"mtime":1606452870087,"results":"8","hashOfConfig":"6"},{"size":197,"mtime":1606452814211,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"zmnk3t",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\App.js",[],"C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\reportWebVitals.js",[],"C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\components\\form\\form.component.jsx",[],"C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\components\\Header\\header.component.jsx",[]]
+[{"C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\reportWebVitals.js":"1","C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\App.js":"2","C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\components\\Form.component.jsx":"3"},{"size":362,"mtime":1606433682700,"results":"4","hashOfConfig":"5"},{"size":254,"mtime":1606456347000,"results":"6","hashOfConfig":"5"},{"size":739,"mtime":1606458319380,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"zmnk3t",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\reportWebVitals.js",[],"C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\App.js",[],"C:\\Users\\Angelie\\Documents\\GitHub\\inventory\\client\\src\\components\\Form.component.jsx",[]]
