--- client/src/components/Header/header.css
@@ -1,4 +1,0 @@
-h1,
-h2 {
-  text-align: center;
-}
