--- client/src/components/form/form.css
@@ -1,21 +1,0 @@
-form {
-  display: flex;
-  flex-direction: column;
-  /* width: 300px; */
-  align-items: center;
-  justify-content: center;
-}
-
-input[type="submit"] {
-  width: 150px;
-  margin-top: 15px;
-}
-
-input {
-  display: block;
-  margin-top: 5px;
-}
-
-label {
-  margin-top: 15px;
-}
